refactor(wayland): separate wayland driver into multiple parts
