+++ src/drivers/wayland/lv_wayland.h
@@ -15,8 +15,6 @@
 #ifndef LV_WAYLAND_H
 #define LV_WAYLAND_H
 
-#ifndef _WIN32
-
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -25,11 +23,15 @@ extern "C" {
  *      INCLUDES
  *********************/
 
-#include "../../display/lv_display.h"
-#include "../../indev/lv_indev.h"
-#include "../../indev/lv_indev_gesture.h"
+#include "../../lv_conf_internal.h"
 
 #if LV_USE_WAYLAND
+
+#include "lv_wl_keyboard.h"
+#include "lv_wl_pointer.h"
+#include "lv_wl_touch.h"
+#include "lv_wl_window.h"
+#include "lv_wl_pointer_axis.h"
 
 /*********************
  *      DEFINES
@@ -39,86 +41,9 @@ extern "C" {
  *      TYPEDEFS
  **********************/
 
-typedef bool (*lv_wayland_display_close_f_t)(lv_display_t * disp);
-
 /**********************
  * GLOBAL PROTOTYPES
  **********************/
-
-/**
- * Retrieves the file descriptor of the wayland socket
- */
-int lv_wayland_get_fd(void);
-
-/**
- * Creates a window
- * @param hor_res The width of the window in pixels
- * @param ver_res The height of the window in pixels
- * @param title The title of the window
- * @param close_cb The callback that will be execute when the user closes the window
- * @return The LVGL display associated to the window
- */
-lv_display_t * lv_wayland_window_create(uint32_t hor_res, uint32_t ver_res, char * title,
-                                        lv_wayland_display_close_f_t close_cb);
-
-/**
- * Closes the window programmatically
- * @param disp Reference to the LVGL display associated to the window
- */
-void lv_wayland_window_close(lv_display_t * disp);
-
-/**
- * Check if the window is open
- * @param disp Reference to the LVGL display associated to the window
- * @return true: The window is open
- */
-bool lv_wayland_window_is_open(lv_display_t * disp);
-
-/**
- * Sets the fullscreen state of the window
- * @param disp Reference to the LVGL display associated to the window
- * @param fullscreen If true the window enters fullscreen
- */
-void lv_wayland_window_set_fullscreen(lv_display_t * disp, bool fullscreen);
-
-/**
- * Sets the maximized state of the window
- * @param disp Reference to the LVGL display associated to the window
- * @param fullscreen If true the window is maximized
- */
-void lv_wayland_window_set_maximized(lv_display_t * disp, bool maximize);
-
-/**
- * Obtains the input device of the mouse pointer
- * @note It is used to create an input group on application start
- * @param disp Reference to the LVGL display associated to the window
- * @return The input device
- */
-lv_indev_t * lv_wayland_get_pointer(lv_display_t * disp);
-
-/**
- * Obtains the input device of the encoder
- * @note It is used to create an input group on application start
- * @param disp Reference to the LVGL display associated to the window
- * @return The input device
- */
-lv_indev_t * lv_wayland_get_pointeraxis(lv_display_t * disp);
-
-/**
- * Obtains the input device of the keyboard
- * @note It is used to create an input group on application start
- * @param disp Reference to the LVGL display associated to the window
- * @return The input device
- */
-lv_indev_t * lv_wayland_get_keyboard(lv_display_t * disp);
-
-/**
- * Obtains the input device of the touch screen
- * @note It is used to create an input group on application start
- * @param disp Reference to the LVGL display associated to the window
- * @return The input device
- */
-lv_indev_t * lv_wayland_get_touchscreen(lv_display_t * disp);
 
 /**
  * Wrapper around lv_timer_handler
@@ -128,6 +53,11 @@ lv_indev_t * lv_wayland_get_touchscreen(lv_display_t * disp);
  * went to sleep because the last frame wasn't completed
  */
 bool lv_wayland_timer_handler(void);
+
+/**
+ * Retrieves the file descriptor of the wayland socket
+ */
+int lv_wayland_get_fd(void);
 
 /**********************
  *      MACROS
@@ -139,5 +69,4 @@ bool lv_wayland_timer_handler(void);
 } /* extern "C" */
 #endif
 
-#endif /* _WIN32 */
 #endif /* WAYLAND_H */
