+++ src/drivers/wayland/lv_wl_buffer.c
@@ -1,0 +1,47 @@
+/**
+ * @file lv_wl_buffer.c
+ *
+ */
+
+/*********************
+ *      INCLUDES
+ *********************/
+#include "lv_wayland.h"
+
+#if LV_USE_WAYLAND
+
+    #include "lv_wayland_private.h"
+
+    /*********************
+    *      DEFINES
+    *********************/
+
+    /**********************
+    *      TYPEDEFS
+    **********************/
+
+    /**********************
+    *  STATIC PROTOTYPES
+    **********************/
+
+    /**********************
+    *  STATIC VARIABLES
+    **********************/
+
+    /**********************
+    *      MACROS
+    **********************/
+
+    /**********************
+    *   GLOBAL FUNCTIONS
+    **********************/
+
+    /**********************
+    *   PRIVATE FUNCTIONS
+    **********************/
+
+    /**********************
+    *   STATIC FUNCTIONS
+    **********************/
+
+#endif /* LV_USE_WAYLAND */
