+++ src/drivers/wayland/lv_wl_keyboard.h
@@ -3,7 +3,6 @@
  *
  */
 
-
 #ifndef LV_WL_KEYBOARD_H
 #define LV_WL_KEYBOARD_H
 
@@ -32,6 +31,13 @@ extern "C" {
  **********************/
 
 lv_indev_t * lv_wayland_keyboard_create(void);
+
+/**
+ * Get keyboard input device for given LVGL display
+ * @param display LVGL display
+ * @return input device connected to keyboard, or NULL on error
+ */
+lv_indev_t * lv_wayland_get_keyboard(lv_display_t * display);
 
 /**********************
  *      MACROS
