+++ src/drivers/wayland/lv_wl_pointer.h
@@ -4,7 +4,6 @@
  *
  */
 
-
 #ifndef LV_WL_POINTER_H
 #define LV_WL_POINTER_H
 
@@ -33,6 +32,14 @@ extern "C" {
  **********************/
 
 lv_indev_t * lv_wayland_pointer_create(void);
+
+/**
+ * Obtains the input device of the mouse pointer
+ * @note It is used to create an input group on application start
+ * @param disp Reference to the LVGL display associated to the window
+ * @return The input device
+ */
+lv_indev_t * lv_wayland_get_pointer(lv_display_t * disp);
 
 /**********************
  *      MACROS
