+++ src/drivers/wayland/lv_wl_pointer_axis.c
@@ -25,7 +25,7 @@
  *  STATIC PROTOTYPES
  **********************/
 
-static void _lv_wayland_pointeraxis_read(lv_indev_t * drv, lv_indev_data_t * data);
+static void pointeraxis_read(lv_indev_t * drv, lv_indev_data_t * data);
 
 /**********************
  *  STATIC VARIABLES
@@ -44,15 +44,25 @@ lv_indev_t * lv_wayland_pointer_axis_create(void)
 
     lv_indev_t * indev = lv_indev_create();
     lv_indev_set_type(indev, LV_INDEV_TYPE_ENCODER);
-    lv_indev_set_read_cb(indev, _lv_wayland_pointeraxis_read);
+    lv_indev_set_read_cb(indev, pointeraxis_read);
 
     return indev;
 }
+
+lv_indev_t * lv_wayland_get_pointeraxis(lv_display_t * display)
+{
+    struct window * window = lv_display_get_user_data(display);
+    if(!window) {
+        return NULL;
+    }
+    return window->lv_indev_pointeraxis;
+}
+
 /**********************
  *   STATIC FUNCTIONS
  **********************/
 
-static void _lv_wayland_pointeraxis_read(lv_indev_t * drv, lv_indev_data_t * data)
+static void pointeraxis_read(lv_indev_t * drv, lv_indev_data_t * data)
 {
     struct window * window = lv_display_get_user_data(lv_indev_get_display(drv));
