+++ src/drivers/wayland/lv_wl_pointer_axis.h
@@ -33,6 +33,14 @@ extern "C" {
 
 lv_indev_t * lv_wayland_pointer_axis_create(void);
 
+/**
+ * Obtains the input device of the encoder
+ * @note It is used to create an input group on application start
+ * @param display Reference to the LVGL display associated to the window
+ * @return The input device
+ */
+lv_indev_t * lv_wayland_get_pointeraxis(lv_display_t * display);
+
 /**********************
  *      MACROS
  **********************/
